requirements.txt
opencv-python
numpy
ultralytics
picamera2
pyserial